New functions

+++ libgmouse/dips.h
@@ -1,6 +1,6 @@
 #ifndef DIPS_H_INCLUDED
 #define DIPS_H_INCLUDED
-uint16_t get_dpi(int8_t dpi_w, int8_t dpi_h);
+unsigned short get_dpi(int dpi_w, int dpi_h);
 
 
 #endif // DIPS_H_INCLUDED
